requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.7
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
filelock==3.0.12
flake8==3.9.2
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
pipenv==2021.5.29
prompt-toolkit==3.0.20
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
wcwidth==0.2.5
